Introduce LanguageClient_hoverMarginSize
The markdown plugin I use has a bunch of regex matching groups that look
for certain characters at the start of a line (like `###`), and the
syntax highlighting breaks when they're not on the start of the line.

We add a single space to the start of every line in a hover comment.
I'd like to make that configurable, so that I can get better markdown
syntax highlighting.

I've chosen to let the user set how big they want the margin to be as an
integer. I'd also be fine with an option that just removes the margin,
but this code is actually shorter to write.